Add DataKinds and AutoDeriveTypeable
On ghc7.8 TH does not support standalone deriving instances.
Additionally, the new Typeable instances can only be derived, not
implemented manually. Thus it is not possible for TH to make tuples
instaces of Typeable.

Instead the language extension "AutoDeriveTypeable" is used to
automatically derive alle needed Typeable instances:

        - Disabled generation of typeable instances for tuples in TH
        - Enabled extensions DataKinds and AutoDeriveTypeable globally

FIXME: Determine the exact scope needed for the extensions and replace
the global option in the .cabal File with specific options in the
corresponding files.