Basic translation of map,foldl,foldr for vectors
This implements the translation of higher order functions for vectors.
There are still some limitations though. A higher-order function (HoF)
application can only be translated, if it is the sole expression of a
function body. This simplification was needed since there is no way of
accessing GHCs type information during TH to VHDL translation, which
means types can only be derived from explicitly written signature
declarations. In case of the HoF translation, these are required within
the body of the translated function.

 - add function specification of the current function to the local
   translation state
 - add HoF translation table
 - add HoF translation functions for map,foldr and foldl
 - some examples for proper usage

*note* as of now, the generated VHDL code is untested.